requirements.txt
streamlit==0.88.0
pandas==1.3.3
plotly==5.3.1
Pillow==8.3.2
numpy 
seaborn 
matplotlib 
